% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-central_atlantic_2_call_area.R
\docType{data}
\name{central_atlantic_2_call_area}
\alias{central_atlantic_2_call_area}
\title{Central Atlantic 2 Call Area}
\source{
\url{https://www.boem.gov/renewable-energy/state-activities/central-atlantic}
}
\usage{
data("central_atlantic_2_call_area")
}
\description{
An \code{sf} object containing the central Atlantic 2 call area.
}
\keyword{datasets}
