% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sea_turtle_critical_habitat.R
\docType{data}
\name{sea_turtle_critical_habitat}
\alias{sea_turtle_critical_habitat}
\title{Sea Turtle Critical Habitat}
\source{
\url{https://www.fisheries.noaa.gov/national/endangered-species-conservation/critical-habitat#critical-habitat-designations-maps-and-gis-data}
}
\usage{
data("sea_turtle_critical_habitat")
}
\description{
An \code{sf} object containing the delineations of sea turtle critical habitat.
}
\keyword{datasets}
